+++ core/src/thing/Thing.ts
@@ -6,6 +6,15 @@ import { isRdfType } from "./isRdfType";
 import { labelForType } from "./labelForType";
 import { labelFromUri } from "./labelFromUri";
 import { Store } from "../Store";
+import {
+  debounceTime,
+  distinctUntilChanged,
+  filter,
+  map,
+  merge,
+  Observable,
+  startWith,
+} from "rxjs";
 
 export interface Literal {
   predicate: string;
@@ -103,6 +112,27 @@ export class Thing {
   }
 
   /**
+   * Observe changes in links from this thing to other resources
+   */
+  observeRelations(predicate?: string): Observable<Relation[]> {
+    return merge(this.store.additions$, this.store.removals$).pipe(
+      // Note: we assume that cost of filtering by the optional predicate is not worthwhile
+      filter((quad) => quad.subject.value == this.uri),
+      debounceTime(250),
+      map(() => this.relations(predicate)),
+      // Note: will not trigger an update if label changes, as label is currently constructed from predicate
+      distinctUntilChanged((prev, curr) =>
+        prev.every(
+          (rel, i) =>
+            rel.predicate == curr[i].predicate &&
+            rel.uris.length == curr[i].uris.length,
+        ),
+      ),
+      startWith(this.relations(predicate)),
+    );
+  }
+
+  /**
    * Returns all the links from other resources to this thing
    */
   reverseRelations(predicate?: string): Relation[] {
@@ -119,6 +149,27 @@ export class Thing {
       label: labelFromUri(predicate),
       uris: values[predicate],
     }));
+  }
+
+  /**
+   * Observe changes in links from other resources to this thing
+   */
+  observeReverseRelations(predicate?: string): Observable<Relation[]> {
+    return merge(this.store.additions$, this.store.removals$).pipe(
+      // Note: we assume that cost of filtering by the optional predicate is not worthwhile
+      filter((quad) => quad.object.value == this.uri),
+      debounceTime(250),
+      map(() => this.reverseRelations(predicate)),
+      // Note: will not trigger an update if label changes, as label is currently constructed from predicate
+      distinctUntilChanged((prev, curr) =>
+        prev.every(
+          (rel, i) =>
+            rel.predicate == curr[i].predicate &&
+            rel.uris.length == curr[i].uris.length,
+        ),
+      ),
+      startWith(this.reverseRelations(predicate)),
+    );
   }
 
   /**
@@ -218,6 +269,25 @@ export class Thing {
       uri,
       label: labelForType(uri),
     }));
+  }
+
+  /**
+   * Observe changes to the list of RDF types for this thing
+   */
+  observeTypes(): Observable<RdfType[]> {
+    return merge(this.store.additions$, this.store.removals$).pipe(
+      filter(
+        (quad) =>
+          (quad.subject.value == this.uri &&
+            quad.predicate.value ==
+              "http://www.w3.org/1999/02/22-rdf-syntax-ns#type") ||
+          quad.predicate.value ==
+            "http://www.w3.org/2000/01/rdf-schema#subClassOf",
+      ),
+      map(() => this.types()),
+      startWith(this.types()),
+      distinctUntilChanged((prev, curr) => prev.length == curr.length),
+    );
   }
 
   /**