+++ core/CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to
 - [`Store.anyValue`]((https://pod-os.org/reference/core/classes/Store/#anyValue)) returns the value of any one RDF/JS term matching the first wildcard in the provided quad pattern
 - [`Store.profileQuery`]((https://pod-os.org/reference/core/classes/Store/#profileQuery)) creates a [query](https://solid-contrib.github.io/data-modules/rdflib-utils/classes/index.ProfileQuery.html) to fetch information from a user's profile document
 - [`Store.preferencesQuery`]((https://pod-os.org/reference/core/classes/Store/#preferencesQuery)) creates a [query](https://solid-contrib.github.io/data-modules/rdflib-utils/classes/index.PreferencesQuery.html) to fetch information from a user's preferences file
+- [`Thing.observeTypes`](https://pod-os.org/reference/core/classes/Thing/#observeTypes) pushes changes to types
+- [`Thing.observeRelations`](https://pod-os.org/reference/core/classes/Thing/#observeRelations) pushes changes to relations
+- [`Thing.observeReverseRelations`](https://pod-os.org/reference/core/classes/Thing/#observeReverseRelations) pushes changes to reverse relations
 
 ## 0.25.0